building fe assets synchronously
