+++ deploy/ec2/.env
@@ -7,6 +7,7 @@ PG_HOST=__required__
 PG_PASS=__required__
 PG_DB=tooljet_prod
 ORM_LOGGING=true
+NODE_ENV=production
 DEPLOYMENT_PLATFORM=ec2
 
 # ToolJet Database
